# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cloud Notes API
  version: 0.1.0
  summary: API contract for the cloud-based notes platform.
servers:
  - url: http://localhost:4000
    description: Local development API
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh and access tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Token rotation successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /folders:
    get:
      tags: [Folders]
      summary: List folders owned by authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Folder list
          content:
            application/json:
              schema:
                type: object
                properties:
                  folders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Folder"
                required: [folders]
    post:
      tags: [Folders]
      summary: Create folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                type: object
                properties:
                  folder:
                    $ref: "#/components/schemas/Folder"
                required: [folder]
  /folders/{folderId}:
    patch:
      tags: [Folders]
      summary: Update folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  folder:
                    $ref: "#/components/schemas/Folder"
                required: [folder]
    delete:
      tags: [Folders]
      summary: Delete folder
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
      responses:
        "204":
          description: Folder deleted
...
  /notes/{noteId}/share/{shareId}:
    patch:
      tags: [Sharing]
      summary: Update share link
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NoteId"
        - $ref: "#/components/parameters/ShareId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShareLinkRequest"
      responses:
        "200":
          description: Share link updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  shareLink:
                    $ref: "#/components/schemas/ShareLink"
                required: [shareLink]
    delete:
      tags: [Sharing]
      summary: Revoke share link
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NoteId"
        - $ref: "#/components/parameters/ShareId"
      responses:
        "204":
          description: Share link revoked
  /shared/{token}:
    get:
      tags: [Sharing]
      summary: Read shared note by public token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared note content
          content:
            application/json:
              schema:
                type: object
                properties:
                  note:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      type:
                        type: string
                        enum: [note, checklist]
                      title:
                        type: string
                      content:
                        type: string
                      metadata:
                        type: object
                        additionalProperties: true
                      updatedAt:
                        type: string
                        format: date-time
                      accessLevel:
                        type: string
                        enum: [view, edit]
                    required: [id, type, title, content, metadata, updatedAt, accessLevel]
                  attachments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Attachment"
                required: [note, attachments]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    NoteId:
      in: path
      name: noteId
      required: true
      schema:
        type: string
        format: uuid
    FolderId:
      in: path
      name: folderId
      required: true
      schema:
        type: string
        format: uuid
    ShareId:
      in: path
      name: shareId
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
      required: [id, email, displayName]
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokens:
          $ref: "#/components/schemas/TokenPair"
      required: [user, tokens]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
      required: [email, password, displayName]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    Folder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ownerId, name, icon, color, createdAt, updatedAt]
    CreateFolderRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
      required: [name]
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
          pattern: "^#[0-9a-fA-F]{6}$"
...
    ShareLink:
      type: object
      properties:
        id:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
        token:
          type: string
        accessLevel:
          type: string
          enum: [view, edit]
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, noteId, token, accessLevel, isActive, expiresAt, createdAt]
...
    UpdateShareLinkRequest:
      type: object
      properties:
        accessLevel:
          type: string
          enum: [view, edit]
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
          enum: [image/png, image/jpeg, image/webp, application/pdf]
        sizeBytes:
          type: integer
          minimum: 1
        publicUrl:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, fileName, mimeType, sizeBytes, publicUrl, createdAt]
...